Reverting PR 42545 (for now) (#44277)
