+++ renovate.json5
@@ -578,14 +578,6 @@
         ],
       },
       {
-        groupSlug: 'archiver',
-        groupName: 'archiver related packages',
-        packageNames: [
-          'archiver',
-          '@types/archiver',
-        ],
-      },
-      {
         groupSlug: 'base64-js',
         groupName: 'base64-js related packages',
         packageNames: [
