+++ x-pack/legacy/plugins/canvas/.storybook/storyshots.test.js
@@ -53,14 +53,6 @@ jest.mock(
   }
 );
 
-// Disabling this test due to https://github.com/elastic/eui/issues/2242
-jest.mock(
-  '../public/components/workpad_header/workpad_export/flyout/__examples__/external_embed_flyout.examples',
-  () => {
-    return 'Disabled Panel';
-  }
-);
-
 addSerializer(styleSheetSerializer);
 
 // Initialize Storyshots and build the Jest Snapshots
