--- x-pack/legacy/plugins/canvas/external_runtime/components/canvas.module.scss
@@ -1,26 +1,0 @@
-:global html body .kbnCanvas {
-  height: auto;
-}
-
-:global .kbnCanvas :local .root {
-  position: relative;
-  overflow: hidden;
-  transition: height 1s;
-}
-
-.container {
-  composes: canvas from global;
-  composes: canvasContainer from global;
-}
-
-:global .kbnCanvas :local .container {
-  align-items: center;
-  display: flex;
-  justify-content: center;
-  pointer-events: none;
-}
-
-:global .kbnCanvas :local .page {
-  position: absolute;
-  transform-origin: center center;
-}
