--- x-pack/legacy/plugins/canvas/external_runtime/components/footer/page_preview.module.scss
@@ -1,15 +1,0 @@
-@import '@elastic/eui/src/global_styling/variables/_size.scss';
-
-:global .kbnCanvas :local .root {
-  margin: 0 $euiSizeS;
-  cursor: pointer;
-}
-
-:global .kbnCanvas :local .root :global .canvasPage {
-  position: relative;
-}
-
-:global .kbnCanvas :local .preview {
-  pointer-events: none;
-  transform-origin: top left;
-}
