--- x-pack/legacy/plugins/canvas/external_runtime/components/footer/settings/index.ts
@@ -1,7 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License;
- * you may not use this file except in compliance with the Elastic License.
- */
-
-export * from './settings';
