--- x-pack/legacy/plugins/canvas/external_runtime/components/rendered_element.module.scss
@@ -1,19 +1,0 @@
-:global .kbnCanvas :local .root,
-:global .kbnCanvas :local .render {
-  height: 100%;
-  width: 100%;
-}
-
-.container {
-  composes: canvas__element from global;
-  composes: canvasElement from global;
-}
-
-.content {
-  composes: canvasElement__content from global;
-}
-
-.renderContainer {
-  composes: canvasWorkpad--element_render from global;
-  composes: canvasRenderEl from global;
-}
