--- x-pack/legacy/plugins/canvas/external_runtime/constants.d.ts
@@ -1,13 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License;
- * you may not use this file except in compliance with the Elastic License.
- */
-
-export const KIBANA_ROOT: string;
-export const LIBRARY_NAME: string;
-export const RUNTIME_FILE: string;
-export const RUNTIME_NAME: string;
-export const RUNTIME_OUTPUT: string;
-export const RUNTIME_SRC: string;
-export const STATS_OUTPUT: string;
