--- x-pack/legacy/plugins/canvas/external_runtime/template.html
@@ -1,14 +1,0 @@
-<html>
-
-<head>
-  <script src="kbn_canvas.js"></script>
-</head>
-
-<body>
-  <div kbn-canvas-embed="canvas" kbn-canvas-height="400" kbn-canvas-url="workpad.json"></div>
-</body>
-<script type="text/javascript">
-  KbnCanvas.embed();
-</script>
-
-</html>
