+++ x-pack/legacy/plugins/canvas/public/components/workpad_header/workpad_export/index.ts
@@ -4,36 +4,24 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-import chrome from 'ui/chrome';
 import { connect } from 'react-redux';
 import { compose, withProps } from 'recompose';
 import { jobCompletionNotifications } from '../../../../../reporting/public/lib/job_completion_notifications';
 // @ts-ignore Untyped local
-import { getWorkpad, getPages, getRenderedWorkpad } from '../../../state/selectors/workpad';
+import { getWorkpad, getPages } from '../../../state/selectors/workpad';
 // @ts-ignore Untyped local
 import { getReportingBrowserType } from '../../../state/selectors/app';
 // @ts-ignore Untyped local
 import { notify } from '../../../lib/notify';
 import { getWindow } from '../../../lib/get_window';
 // @ts-ignore Untyped local
-import {
-  downloadWorkpad,
-  downloadRenderedWorkpad,
-  downloadEmbedRuntime,
-  downloadZippedEmbed,
-  // @ts-ignore Untyped local
-} from '../../../lib/download_workpad';
+import { downloadWorkpad } from '../../../lib/download_workpad';
 import { WorkpadExport as Component, Props as ComponentProps } from './workpad_export';
 import { getPdfUrl, createPdf } from './utils';
 import { CanvasWorkpad } from '../../../../types';
-import { CanvasRenderedWorkpad } from '../../../../external_runtime/types';
-// @ts-ignore Untyped local.
-import { fetch, arrayBufferFetch } from '../../../../common/lib/fetch';
-import { API_ROUTE_SNAPSHOT_ZIP } from '../../../../common/lib/constants';
 
 const mapStateToProps = (state: any) => ({
   workpad: getWorkpad(state),
-  renderedWorkpad: getRenderedWorkpad(state),
   pageCount: getPages(state).length,
   enabled: getReportingBrowserType(state) === 'chromium',
 });
@@ -51,7 +39,6 @@ const getAbsoluteUrl = (path: string) => {
 
 interface Props {
   workpad: CanvasWorkpad;
-  renderedWorkpad: CanvasRenderedWorkpad;
   pageCount: number;
   enabled: boolean;
 }
@@ -59,7 +46,7 @@ interface Props {
 export const WorkpadExport = compose<ComponentProps, Props>(
   connect(mapStateToProps),
   withProps(
-    ({ workpad, pageCount, enabled, renderedWorkpad }: Props): ComponentProps => ({
+    ({ workpad, pageCount, enabled }: Props): ComponentProps => ({
       enabled,
       getExportUrl: type => {
         if (type === 'pdf') {
@@ -76,11 +63,8 @@ export const WorkpadExport = compose<ComponentProps, Props>(
           case 'reportingConfig':
             notify.info(`Copied reporting configuration to clipboard`);
             break;
-          case 'embed':
-            notify.info(`Copied embed code to clipboard`);
-            break;
           default:
-            throw new Error(`Unknown copy type: ${type}`);
+            throw new Error(`Unknown export type: ${type}`);
         }
       },
       onExport: type => {
@@ -100,19 +84,7 @@ export const WorkpadExport = compose<ComponentProps, Props>(
               });
           case 'json':
             downloadWorkpad(workpad.id);
-            return;
-          case 'embed':
-            downloadRenderedWorkpad(renderedWorkpad);
-            return;
-          case 'runtime':
-            downloadEmbedRuntime();
-            return;
-          case 'zip':
-            const basePath = chrome.getBasePath();
-            arrayBufferFetch
-              .post(`${basePath}${API_ROUTE_SNAPSHOT_ZIP}`, JSON.stringify(renderedWorkpad))
-              .then(blob => downloadZippedEmbed(blob.data));
-            return;
+            break;
           default:
             throw new Error(`Unknown export type: ${type}`);
         }
