+++ x-pack/legacy/plugins/canvas/public/lib/download_workpad.js
@@ -4,8 +4,6 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 import fileSaver from 'file-saver';
-import chrome from 'ui/chrome';
-import { API_ROUTE_SNAPSHOT_RUNTIME_DOWNLOAD } from '../../common/lib/constants';
 import { notify } from './notify';
 import * as workpadService from './workpad_service';
 
@@ -16,37 +14,5 @@ export const downloadWorkpad = async workpadId => {
     fileSaver.saveAs(jsonBlob, `canvas-workpad-${workpad.name}-${workpad.id}.json`);
   } catch (err) {
     notify.error(err, { title: `Couldn't download workpad` });
-  }
-};
-
-export const downloadRenderedWorkpad = async renderedWorkpad => {
-  try {
-    const jsonBlob = new Blob([JSON.stringify(renderedWorkpad)], { type: 'application/json' });
-    fileSaver.saveAs(
-      jsonBlob,
-      `canvas-embed-workpad-${renderedWorkpad.name}-${renderedWorkpad.id}.json`
-    );
-  } catch (err) {
-    notify.error(err, { title: `Couldn't download rendered workpad` });
-  }
-};
-
-export const downloadEmbedRuntime = async () => {
-  try {
-    const basePath = chrome.getBasePath();
-    const path = `${basePath}${API_ROUTE_SNAPSHOT_RUNTIME_DOWNLOAD}`;
-    window.open(path);
-    return;
-  } catch (err) {
-    notify.error(err, { title: `Couldn't download embed runtime` });
-  }
-};
-
-export const downloadZippedEmbed = async data => {
-  try {
-    const zip = new Blob([data], { type: 'octet/stream' });
-    fileSaver.saveAs(zip, 'canvas-workpad-embed.zip');
-  } catch (err) {
-    notify.error(err, { title: `Couldn't download ZIP file` });
   }
 };
