+++ x-pack/package.json
@@ -44,7 +44,6 @@
     "@storybook/react": "^5.0.5",
     "@storybook/theming": "^5.0.5",
     "@types/angular": "1.6.50",
-    "@types/archiver": "^3.0.0",
     "@types/base64-js": "^1.2.5",
     "@types/boom": "^7.2.0",
     "@types/cheerio": "^0.22.10",
@@ -152,7 +151,6 @@
     "mustache": "^2.3.0",
     "mutation-observer": "^1.0.3",
     "node-fetch": "^2.6.0",
-    "null-loader": "^3.0.0",
     "pdf-image": "2.0.0",
     "pdfjs-dist": "^2.0.943",
     "pixelmatch": "4.0.2",
@@ -184,7 +182,6 @@
     "@babel/runtime": "^7.5.5",
     "@elastic/ctags-langserver": "^0.1.8",
     "@elastic/datemath": "5.0.2",
-    "@elastic/filesaver": "1.1.2",
     "@elastic/eui": "13.6.1",
     "@elastic/javascript-typescript-langserver": "^0.2.2",
     "@elastic/lsp-extension": "^0.1.2",
@@ -219,7 +216,6 @@
     "apollo-link-state": "^0.4.1",
     "apollo-server-errors": "^2.0.2",
     "apollo-server-hapi": "^1.3.6",
-    "archiver": "3.1.1",
     "axios": "^0.19.0",
     "bluebird": "3.5.5",
     "boom": "^7.2.0",
@@ -305,7 +301,6 @@
     "pngjs": "3.4.0",
     "polished": "^1.9.2",
     "popper.js": "^1.14.3",
-    "postcss-prefix-selector": "^1.7.2",
     "prop-types": "^15.6.0",
     "proper-lockfile": "^3.2.0",
     "puid": "1.0.7",
@@ -366,7 +361,6 @@
     "uuid": "3.3.2",
     "venn.js": "0.2.20",
     "vscode-languageserver": "^5.2.1",
-    "webpack": "4.33.0",
     "wellknown": "^0.5.0",
     "xml2js": "^0.4.19",
     "xregexp": "4.2.4"
